UPDATE - Updated middlewares
